Get rid of the Term type for elections.
Just to be clear - I think type aliases are great. If you're going to
make a type alias, though, please use it consistently. This type alias
was only used in a few places (mostly tests), while the type it was
aliasing (u64) was used in the actual application code. This just leads
to confusion.

Signed-off-by: Josh Black <raskchanky@gmail.com>